Move graph to own comp file - #662
